+++ pkg/analyzer/endpoints/ipBlock.go
@@ -11,22 +11,26 @@ type IPBlock struct {
 }
 
 func (b *IPBlock) Name() string {
+	// todo
 	return b.name
 }
 
 func (b *IPBlock) String() string {
+	// todo
 	return b.Name()
 }
 
 func (b *IPBlock) ID() string {
+	// todo
 	return b.Name()
 }
 
 func (b *IPBlock) Kind() string {
-	return "external"
+	return "block"
 }
 
 func (b *IPBlock) InfoStr() []string {
+	// todo
 	return []string{b.Name(), b.ID(), strings.Join([]string{b.Name()}, common.CommaSeparator)}
 }
